test(select): add shape snapshots
